Drop for the character done

+++ server_files/server_helmet_factory.cpp
@@ -4,7 +4,7 @@
 #define IRON_HELMET 5
 #define MAGIC_HAT 6
 
-Helmet HelmetFactory::make_helmet(const int helmet_id, Json::Value& config) {
+Helmet HelmetFactory::make_helmet(const uint8_t helmet_id, Json::Value& config) {
     switch (helmet_id) {
         case HOOD: return Hood(config);
         case IRON_HELMET: return IronHelmet(config);
